Refactor SVGMobject using svgelements
