+++ setup.cfg
@@ -50,7 +50,6 @@ install_requires =
     ipython
     PyOpenGL
     manimpango>=0.2.0,<0.4.0
-    cssselect2
     isosurfaces
     svgelements
